Refactored the overriden functions in Antlr parser to a separate class. Added logic and test to map precision of timestamp datatypes.
